finished

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,12 +1,20 @@
 import React from "react";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 import useWorkoutContext from "../hooks/useWorkoutContext";
+import useAuthContext from "../hooks/useAuthContext";
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutContext();
+  const { user } = useAuthContext();
 
   const handleDelete = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
